requirements.txt
# 🤖 وابستگی‌های ربات استودیو ماندنی
# Mandani Studio Bot Dependencies - Production Ready

# Core Telegram Bot Framework
python-telegram-bot==20.7

# PDF Generation
reportlab==4.0.5

# Environment Variables Management
python-dotenv==1.0.0

# Persian Date Support  
jdatetime==4.1.0

# Enhanced Logging
python-json-logger==2.0.7

# HTTP Requests
requests==2.31.0
aiohttp==3.8.6
httpx==0.25.2

# Data Validation
pydantic==2.5.0

# Performance Monitoring
psutil==5.9.6

# Error Tracking (Optional)
sentry-sdk==1.32.0

# Job Scheduling for Reminders
APScheduler==3.10.4

# Additional dependencies for hosting
gunicorn==21.2.0
